+++ src/backend/controllers/AuthController.js
@@ -16,46 +16,48 @@ const jwt = require("jsonwebtoken");
  * */
 
 export const signupHandler = function (schema, request) {
-  const { email, password, ...rest } = JSON.parse(request.requestBody);
-  try {
-    // check if email already exists
-    const foundUser = schema.users.findBy({ email });
-    if (foundUser) {
-      return new Response(
-        422,
-        {},
-        {
-          errors: ["Unprocessable Entity. Email Already Exists."],
-        }
-      );
-    }
-    const _id = uuid();
-    const encryptedPassword = bcrypt.hashSync(password, 5);
-    const newUser = {
-      _id,
-      email,
-      password: encryptedPassword,
-      createdAt: formatDate(),
-      updatedAt: formatDate(),
-      ...rest,
-      cart: [],
-      wishlist: [],
-    };
-    const createdUser = schema.users.create(newUser);
-    const encodedToken = jwt.sign(
-      { _id, email },
-      process.env.REACT_APP_JWT_SECRET
-    );
-    return new Response(201, {}, { createdUser, encodedToken });
-  } catch (error) {
-    return new Response(
-      500,
-      {},
-      {
-        error,
-      }
-    );
-  }
+	const { email, password, ...rest } = JSON.parse(request.requestBody);
+	try {
+		// check if email already exists
+		const foundUser = schema.users.findBy({ email });
+		if (foundUser) {
+			return new Response(
+				422,
+				{},
+				{
+					errors: ["Unprocessable Entity. Email Already Exists."],
+				}
+			);
+		}
+		const _id = uuid();
+		const encryptedPassword = bcrypt.hashSync(password, 5);
+		const newUser = {
+			_id,
+			email,
+			password: encryptedPassword,
+			createdAt: formatDate(),
+			updatedAt: formatDate(),
+			...rest,
+			cart: [],
+			wishlist: [],
+			addressList: [],
+			orders: [],
+		};
+		const createdUser = schema.users.create(newUser);
+		const encodedToken = jwt.sign(
+			{ _id, email },
+			process.env.REACT_APP_JWT_SECRET
+		);
+		return new Response(201, {}, { createdUser, encodedToken });
+	} catch (error) {
+		return new Response(
+			500,
+			{},
+			{
+				error,
+			}
+		);
+	}
 };
 
 /**
@@ -65,40 +67,40 @@ export const signupHandler = function (schema, request) {
  * */
 
 export const loginHandler = function (schema, request) {
-  const { email, password } = JSON.parse(request.requestBody);
-  try {
-    const foundUser = schema.users.findBy({ email });
-    if (!foundUser) {
-      return new Response(
-        404,
-        {},
-        { errors: ["The email you entered is not Registered. Not Found error"] }
-      );
-    }
-    if (bcrypt.compareSync(password, foundUser.password)) {
-      const encodedToken = jwt.sign(
-        { _id: foundUser._id, email },
-        process.env.REACT_APP_JWT_SECRET
-      );
-      foundUser.password = undefined;
-      return new Response(200, {}, { foundUser, encodedToken });
-    }
-    new Response(
-      401,
-      {},
-      {
-        errors: [
-          "The credentials you entered are invalid. Unauthorized access error.",
-        ],
-      }
-    );
-  } catch (error) {
-    return new Response(
-      500,
-      {},
-      {
-        error,
-      }
-    );
-  }
+	const { email, password } = JSON.parse(request.requestBody);
+	try {
+		const foundUser = schema.users.findBy({ email });
+		if (!foundUser) {
+			return new Response(
+				404,
+				{},
+				{ errors: ["The email you entered is not Registered. Not Found error"] }
+			);
+		}
+		if (bcrypt.compareSync(password, foundUser.password)) {
+			const encodedToken = jwt.sign(
+				{ _id: foundUser._id, email },
+				process.env.REACT_APP_JWT_SECRET
+			);
+			foundUser.password = undefined;
+			return new Response(200, {}, { foundUser, encodedToken });
+		}
+		new Response(
+			401,
+			{},
+			{
+				errors: [
+					"The credentials you entered are invalid. Unauthorized access error.",
+				],
+			}
+		);
+	} catch (error) {
+		return new Response(
+			500,
+			{},
+			{
+				error,
+			}
+		);
+	}
 };
