+++ src/backend/controllers/CartController.js
@@ -13,6 +13,7 @@ import { formatDate, requiresAuth } from "../utils/authUtils";
  * */
 export const getCartItemsHandler = function (schema, request) {
 	const userId = requiresAuth.call(this, request);
+	console.log(userId);
 	if (!userId) {
 		new Response(
 			404,
